refactor(cell): split world.rs (1081 lines) into directory module
Extract test module into sibling file:

- mod.rs: world handler logic (616 lines)
- tests.rs: unit tests (465 lines)
