+++ crates/services/src/cell/cell_methods/player/world/mod.rs
@@ -1,0 +1,418 @@
+use crate::cell::client_methods::{being, spawnable_entity};
+use crate::cell::messages::CellToBaseMsg;
+use crate::cell::space_manager::SpaceManager;
+use cimmeria_content_engine::chain::ChainEngine;
+use tokio::sync::mpsc;
+
+use super::constants::*;
+
+pub async fn dispatch(
+    entity_id: u32,
+    method_index: u16,
+    args: &[u8],
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+    engine: &ChainEngine,
+) -> bool {
+    match method_index {
+        SET_AUTO_CYCLE => {
+            if !args.is_empty() {
+                let enabled = args[0] != 0;
+                tracing::debug!(entity_id, enabled, "setAutoCycle");
+                if let Some(entity) = space_mgr.get_entity_mut(entity_id) {
+                    entity.abilities.auto_cycle = enabled;
+                    if !enabled {
+                        entity.abilities.auto_cycle_ability_id = None;
+                    }
+                }
+            }
+            true
+        }
+
+        LOOT_ITEM => {
+            if args.len() >= 4 {
+                let index = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
+                crate::cell::interactions::handle_loot_item(entity_id, index, tx, space_mgr).await;
+            }
+            true
+        }
+
+        TRIGGER_REGION => {
+            if args.len() >= 17 {
+                let region_id = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
+                let b_entering = args[4] != 0;
+                let _x = f32::from_le_bytes([args[5], args[6], args[7], args[8]]);
+                let _y = f32::from_le_bytes([args[9], args[10], args[11], args[12]]);
+                let _z = f32::from_le_bytes([args[13], args[14], args[15], args[16]]);
+
+                // Region IDs are wire-encoded as i32 but stored as u32 internally.
+                // Reject negative values up-front rather than sign-extending them
+                // into a high u32 that no real region will match.
+                let (region_tag, db_set_id) = match u32::try_from(region_id) {
+                    Ok(rid) => match space_mgr.get_region(rid) {
+                        Some(r) => (Some(r.tag.clone()), Some(r.db_set_id)),
+                        None => (None, None),
+                    },
+                    Err(_) => {
+                        tracing::warn!(
+                            entity_id,
+                            region_id,
+                            "triggerClientHintedGenericRegion: negative region_id, ignoring"
+                        );
+                        (None, None)
+                    }
+                };
+
+                if let Some(tag) = region_tag {
+                    tracing::info!(entity_id, region_id, %tag, b_entering, "triggerClientHintedGenericRegion");
+
+                    let player_id = space_mgr
+                        .get_entity(entity_id)
+                        .and_then(|e| e.player_id)
+                        .unwrap_or(0);
+
+                    if b_entering {
+                        crate::cell::content::fire_enter_region(
+                            entity_id, player_id, &tag, engine, tx, space_mgr,
+                        )
+                        .await;
+                    } else {
+                        crate::cell::content::fire_exit_region(
+                            entity_id, player_id, &tag, engine, tx, space_mgr,
+                        )
+                        .await;
+                    }
+
+                    // Forward to the ring transporter FSM if this region is a
+                    // ring pad (point_set_id matches a loaded ring region).
+                    if let Some(set_id) = db_set_id {
+                        crate::cell::ring_transport::handle_region_trigger(
+                            set_id, b_entering, entity_id, tx, space_mgr, engine,
+                        )
+                        .await;
+                    }
+                } else {
+                    tracing::warn!(
+                        entity_id,
+                        region_id,
+                        "Unknown region ID in triggerClientHintedGenericRegion"
+                    );
+                }
+            }
+            true
+        }
+
+        REQUEST_RELOAD => {
+            if !args.is_empty() {
+                let _reload_type = args[0];
+                tracing::debug!(entity_id, "requestReload");
+                handle_reload(entity_id, tx, space_mgr).await;
+            }
+            true
+        }
+
+        CHOSEN_REWARDS => {
+            tracing::info!(entity_id, "UNIMPLEMENTED: chosenRewards");
+            true
+        }
+
+        SET_RING_TRANSPORTER_DEST => {
+            if args.len() >= 8 {
+                let region_id = i32::from_le_bytes([args[0], args[1], args[2], args[3]]);
+                let destination_id = i32::from_le_bytes([args[4], args[5], args[6], args[7]]);
+                tracing::info!(
+                    entity_id,
+                    region_id,
+                    destination_id,
+                    "setRingTransporterDestination"
+                );
+                crate::cell::ring_transport::handle_select_destination(
+                    region_id,
+                    destination_id,
+                    entity_id,
+                    tx,
+                    space_mgr,
+                    engine,
+                )
+                .await;
+            }
+            true
+        }
+
+        WORLD_INSTANCE_RESET => {
+            tracing::info!(entity_id, "UNIMPLEMENTED: onWorldInstanceReset");
+            true
+        }
+
+        UPDATE_SYSTEM_OPTIONS => {
+            tracing::info!(entity_id, "UNIMPLEMENTED: updateSystemOptions");
+            true
+        }
+
+        _ => false,
+    }
+}
+
+const ABILITY_RELOAD_WEAPON: i32 = 596;
+
+/// How long the draw animation needs to play before the reload sequence
+/// can fire. Hand needs to reach the hold position and grip the weapon
+/// mesh; firing `Item_Reload` while the hand is mid-air mid-draw plays
+/// the reload animation on a model that isn't in the reload-ready pose
+/// — the client either ignores the request or visually skips the
+/// reload anim (the symptom the user reported in playtest: "weapon
+/// teleports into my hand and I still need to hit reload again").
+///
+/// Empirically tuned to 1 second; matches the rough length of the
+/// `KIS-handling` kismet script's draw branch. Bump if the reload still
+/// chains into the draw mid-animation; lower if the gap between draw
+/// and reload becomes visually obvious.
+pub(crate) const UNHOLSTER_DRAW_DURATION: std::time::Duration =
+    std::time::Duration::from_millis(1000);
+
+/// Fire a `Item_*` kismet sequence (`Item_Equip` 4000 / `Item_Unequip`
+/// 4001 / `Item_Reload` 4002 / `Item_Use` 4003) keyed off the player's
+/// archetype-keyed "Item handling" event set. Mirrors
+/// `python/cell/SGWBeing.py:getItemSequence(eventId)` + `playSequence`.
+///
+/// No-op (with a debug log) when the archetype, the event set, or the
+/// per-event sequence is missing — matches the python's `if eventSet
+/// else None` fallthrough so callers don't crash on edge entities.
+pub(crate) async fn fire_item_sequence(
+    entity_id: u32,
+    event_id: i32,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &SpaceManager,
+) {
+    let archetype_id = space_mgr.get_entity(entity_id).and_then(|e| e.archetype_id);
+    let event_set = archetype_id.and_then(crate::cell::spawner::archetype_item_event_set);
+    let seq_id = event_set.and_then(|esid| space_mgr.sequence_map.get(&(esid, event_id)).copied());
+    tracing::info!(
+        entity_id,
+        event_id,
+        archetype_id = ?archetype_id,
+        event_set_id = ?event_set,
+        seq_id = ?seq_id,
+        "fire_item_sequence: archetype-keyed sequence lookup"
+    );
+    let Some(seq_id) = seq_id else {
+        return;
+    };
+    // ON_SEQUENCE wire layout (26 bytes — matches use_ability.rs's fire
+    // path so animations are emitted consistently with weapon-fire and
+    // reload animations).
+    let mut seq_args = Vec::with_capacity(28);
+    seq_args.extend_from_slice(&seq_id.to_le_bytes());
+    seq_args.extend_from_slice(&(entity_id as i32).to_le_bytes());
+    seq_args.extend_from_slice(&(entity_id as i32).to_le_bytes());
+    seq_args.push(1);
+    seq_args.extend_from_slice(&0.0f32.to_le_bytes());
+    seq_args.extend_from_slice(&0u32.to_le_bytes());
+    seq_args.push(0);
+    seq_args.extend_from_slice(&0i32.to_le_bytes());
+    let _ = tx
+        .send(CellToBaseMsg::EntityMethodCall {
+            entity_id,
+            method_index: spawnable_entity::ON_SEQUENCE,
+            args: seq_args,
+        })
+        .await;
+}
+
+pub(crate) async fn handle_reload(
+    entity_id: u32,
+    tx: &mpsc::Sender<CellToBaseMsg>,
+    space_mgr: &mut SpaceManager,
+) {
+    let reload_def = space_mgr.ability_defs.get(&ABILITY_RELOAD_WEAPON).cloned();
+    let warmup = reload_def.as_ref().map_or(2.0f32, |d| d.warmup);
+    let cooldown = reload_def.as_ref().map_or(1.0f32, |d| d.cooldown);
+
+    // Phase A — reload-while-holstered: defer the actual reload until
+    // the draw animation has had time to play. Fires `Item_Equip`
+    // (event 4000), the bandolier-equip animation, as a stand-in for an
+    // explicit draw sequence (the 2009 client never shipped one — the
+    // archaeology agent confirmed `Event_NetOut_ChangeWeaponState` is
+    // dead scaffolding). Re-running `handle_reload` after
+    // `UNHOLSTER_DRAW_DURATION` (via `pending_reload_tick`) lands us in
+    // Phase B with the weapon already drawn.
+    //
+    // Gate: weapon currently holstered + threatened_mobs empty (OOC) +
+    // no pending phase already in flight. In-combat reload skips this
+    // entirely (weapon's already drawn).
+    let needs_phase_a = match space_mgr.get_entity(entity_id) {
+        Some(e) => {
+            e.weapon_holstered && e.threatened_mobs.is_empty() && e.pending_reload_at.is_none()
+        }
+        None => false,
+    };
+    if needs_phase_a {
+        // Don't accidentally start a phase A for a player whose mag is
+        // already full — the early-return below would catch it after
+        // Phase B too, but skipping the wasted draw animation is the
+        // right move.
+        let ammo_already_full = match space_mgr.get_entity(entity_id) {
+            Some(e) => e.active_ammo() >= e.active_clip_size() && e.reload_complete_at.is_none(),
+            None => true,
+        };
+        if !ammo_already_full {
+            if let Some(e) = space_mgr.get_entity_mut(entity_id) {
+                e.combat_exit_at = Some(std::time::Instant::now());
+                e.set_weapon_holstered(false);
+                e.pending_reload_at = Some(std::time::Instant::now() + UNHOLSTER_DRAW_DURATION);
+                // Cancel any in-flight holster Phase 2 — reload draws
+                // the weapon BACK out, so a stale Phase 2 would yank
+                // the mesh away mid-reload.
+                e.holster_animation_complete_at = None;
+            }
+            tracing::info!(
+                entity_id,
+                draw_duration_ms = UNHOLSTER_DRAW_DURATION.as_millis() as u64,
+                "reload-while-holstered: phase A — drawing weapon, reload deferred"
+            );
+            crate::cell::abilities::request_appearance_refresh(entity_id, tx, space_mgr).await;
+            fire_item_sequence(
+                entity_id,
+                crate::cell::spawner::EVENT_ITEM_EQUIP,
+                tx,
+                space_mgr,
+            )
+            .await;
+            return;
+        }
+    }
+
+    // Reject second-press during the Phase A draw window. If
+    // `pending_reload_at` is set and the timestamp hasn't elapsed yet,
+    // the only legitimate entry path is the tick — but the tick fires
+    // strictly after the timestamp, so a `now < pending_reload_at`
+    // observation here means the player pressed R again mid-draw.
+    // Without this gate, the second press falls through to Phase B,
+    // clears `pending_reload_at` ahead of schedule, and starts the
+    // reload cooldown immediately — defeating the draw window.
+    if let Some(t) = space_mgr
+        .get_entity(entity_id)
+        .and_then(|e| e.pending_reload_at)
+    {
+        if std::time::Instant::now() < t {
+            tracing::debug!(
+                entity_id,
+                "requestReload: ignoring while draw window in progress"
+            );
+            return;
+        }
+    }
+
+    // Phase B (or a normal already-drawn reload). When entered from the
+    // `pending_reload_tick`, clear the deferred-reload stamp so a
+    // racing tick won't re-fire phase B.
+    if let Some(e) = space_mgr.get_entity_mut(entity_id) {
+        e.pending_reload_at = None;
+    }
+
+    let entity = match space_mgr.get_entity_mut(entity_id) {
+        Some(e) => e,
+        None => {
+            tracing::warn!(entity_id, "requestReload: entity not found");
+            return;
+        }
+    };
+
+    if entity.active_ammo() >= entity.active_clip_size() && entity.reload_complete_at.is_none() {
+        tracing::debug!(entity_id, "requestReload: already at max ammo");
+        return;
+    }
+
+    let old = entity.active_ammo();
+    let target_ammo = entity.active_clip_size();
+
+    let total_time = warmup + cooldown;
+    entity.abilities.start_ability_cooldown(
+        ABILITY_RELOAD_WEAPON,
+        std::time::Duration::from_secs_f32(total_time),
+    );
+
+    // Defer the actual ammo refill until after the warmup. The reload-completion
+    // tick promotes pending refills; the fire-path gates on `reload_complete_at`
+    // to prevent shooting during the warmup.
+    //
+    // Pin the reload to the slot that started it. If the player swaps weapons
+    // mid-reload, the tick must refill *this* slot — not whatever slot is
+    // active when the deadline elapses.
+    let warmup_duration = std::time::Duration::from_secs_f32(warmup.max(0.0));
+    entity.reload_complete_at = Some(std::time::Instant::now() + warmup_duration);
+    entity.reload_slot_id = Some(entity.active_bandolier_slot);
+
+    tracing::info!(
+        entity_id,
+        old,
+        target = target_ammo,
+        warmup,
+        cooldown,
+        "Weapon reload started"
+    );
+
+    let timer_args = cimmeria_entity::abilities::serialize_timer_update(
+        ABILITY_RELOAD_WEAPON,
+        cimmeria_entity::abilities::TIMER_ABILITY_COOLDOWN,
+        entity_id as i32,
+        total_time,
+        0.0,
+    );
+    let _ = tx
+        .send(CellToBaseMsg::EntityMethodCall {
+            entity_id,
+            method_index: being::ON_TIMER_UPDATE,
+            args: timer_args,
+        })
+        .await;
+
+    // BSF_InCombat is intentionally not touched here: a reload in
+    // isolation (no aggro) must not flip the in-combat HUD/cursor. The
+    // bit is derived from `threatened_mobs` and only flips on via
+    // `combat::generate_threat` → `enter_player_combat` when the player
+    // actually generates threat on a surviving NPC.
+
+    // Re-stamp `combat_exit_at` so the OOC holster timer fires
+    // `OOC_HOLSTER_DELAY` seconds from reload start, never mid-animation.
+    // (Phase A already stamped this; Phase B re-stamps so a normal
+    // already-drawn reload also resets the OOC countdown.) In-combat
+    // reload is untouched — `threatened_mobs.is_empty()` is false, so
+    // we skip entirely and `combat_exit_at` stays None until the fight
+    // ends naturally.
+    if let Some(e) = space_mgr.get_entity_mut(entity_id) {
+        if e.threatened_mobs.is_empty() {
+            e.combat_exit_at = Some(std::time::Instant::now());
+        }
+    }
+
+    // Fire the `Item_Reload` (event 4002) animation — the visible
+    // drop-mag / insert-mag / chamber sequence. Mirrors
+    // `python/cell/SGWBeing.py:863-874`'s `getItemSequence` +
+    // `playSequence`. Archetype lookup + sequence dispatch lives in
+    // `fire_item_sequence`.
+    fire_item_sequence(
+        entity_id,
+        crate::cell::spawner::EVENT_ITEM_RELOAD,
+        tx,
+        space_mgr,
+    )
+    .await;
+
+    let mut args = Vec::with_capacity(8);
+    args.extend_from_slice(&7i32.to_le_bytes());
+    let ammo_type = space_mgr
+        .get_entity(entity_id)
+        .map_or(0, |e| e.active_ammo_type());
+    args.extend_from_slice(&ammo_type.to_le_bytes());
+    let _ = tx
+        .send(CellToBaseMsg::EntityMethodCall {
+            entity_id,
+            method_index: spawnable_entity::ON_ENTITY_PROPERTY,
+            args,
+        })
+        .await;
+}
+
+#[cfg(test)]
+mod tests;
